Re-added the files that were deleted in the previous commit
Chose File -> New, then chose to add a Forms ContentPage XAML.
After typing in the same file names and clicking OK, chose "Add existing" in the dialogue that says a file with the same file name already exists in the folder.

This will add back the XAML and the code-behind files that were deleted in the previous commit.